backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg[binary]==3.2.12
python-dotenv==1.0.0
Werkzeug==2.3.0
PyJWT==2.6.0
